Seach Indicator

A function that searches all available indicators based on description input or code. If both are given then parameter code will take preference

search_indicator(description, time_series, code = NULL)

Arguments

description
description of indicator to look up

time_series
time series to look for

code
code of indicator to return meta data from

Value

data.frame

Examples

search_indicator(description = "GDP")

#> # A tibble: 101 x 7
#>    time_series description_and_version         series version_descrip… frequency
#>    <chr>       <glue>                          <chr>  <chr>            <chr>    
#>  1 KBP4116     Total loan debt of national go… F      NA               J2       
#>  2 KBP4116     Total loan debt of national go… J      NA               J1       
#>  3 KBP4116     Total loan debt of national go… K      NA               K1       
#>  4 KBP4117     Total loan debt of national go… F      NA               J2       
#>  5 KBP4117     Total loan debt of national go… J      NA               J1       
#>  6 KBP4117     Total loan debt of national go… K      NA               K1       
#>  7 KBP4419     Primary balance as percentage … F      NA               J2       
#>  8 KBP4419     Primary balance as percentage … J      NA               J1       
#>  9 KBP4419     Primary balance as percentage … K      NA               K1       
#> 10 KBP4420     National government deficit / … F      NA               J2       
#> # … with 91 more rows, and 2 more variables: unit_of_measure <chr>, code <glue>

search_indicator(time_series = "KBP1000")

#> # A tibble: 2 x 7
#>   time_series description_and… series version_descrip… frequency unit_of_measure
#>   <chr>       <glue>           <chr>  <chr>            <chr>     <chr>          
#> 1 KBP1000     South African R… J      NA               J1        RMILL          
#> 2 KBP1000     South African R… M      NA               M1        RMILL          
#> # … with 1 more variable: code <glue>

search_indicator(code = "KBP1000J")

#> # A tibble: 1 x 7
#>   time_series description_and… series version_descrip… frequency unit_of_measure
#>   <chr>       <glue>           <chr>  <chr>            <chr>     <chr>          
#> 1 KBP1000     South African R… J      NA               J1        RMILL          
#> # … with 1 more variable: code <glue>