Seach Indicator

A function that searches all available indicators based on description input or code. If both are given then parameter code will take preference

search_indicator(description, time_series, code = NULL)

Arguments

description
description of indicator to look up

time_series
time series to look for

code
code of indicator to return meta data from

Value

data.frame

Examples

search_indicator(description = "GDP")
#> # A tibble: 101 × 7
#>    time_series description_and_version      series versi…¹ frequ…² unit_…³ code 
#>    <chr>       <glue>                       <chr>  <chr>   <chr>   <chr>   <glu>
#>  1 KBP4116     Total loan debt of national… F      NA      J2      RMILL   KBP4…
#>  2 KBP4116     Total loan debt of national… J      NA      J1      RMILL   KBP4…
#>  3 KBP4116     Total loan debt of national… K      NA      K1      RMILL   KBP4…
#>  4 KBP4117     Total loan debt of national… F      NA      J2      RMILL   KBP4…
#>  5 KBP4117     Total loan debt of national… J      NA      J1      RMILL   KBP4…
#>  6 KBP4117     Total loan debt of national… K      NA      K1      RMILL   KBP4…
#>  7 KBP4419     Primary balance as percenta… F      NA      J2      PERC    KBP4…
#>  8 KBP4419     Primary balance as percenta… J      NA      J1      PERC    KBP4…
#>  9 KBP4419     Primary balance as percenta… K      NA      K1      PERC    KBP4…
#> 10 KBP4420     National government deficit… F      NA      J2      PERC    KBP4…
#> # … with 91 more rows, and abbreviated variable names ¹​version_description,
#> #   ²​frequency, ³​unit_of_measure
search_indicator(time_series = "KBP1000")
#> # A tibble: 2 × 7
#>   time_series description_and_version       series versi…¹ frequ…² unit_…³ code 
#>   <chr>       <glue>                        <chr>  <chr>   <chr>   <chr>   <glu>
#> 1 KBP1000     South African Reserve Bank l… J      NA      J1      RMILL   KBP1…
#> 2 KBP1000     South African Reserve Bank l… M      NA      M1      RMILL   KBP1…
#> # … with abbreviated variable names ¹​version_description, ²​frequency,
#> #   ³​unit_of_measure
search_indicator(code = "KBP1000J")
#> # A tibble: 1 × 7
#>   time_series description_and_version       series versi…¹ frequ…² unit_…³ code 
#>   <chr>       <glue>                        <chr>  <chr>   <chr>   <chr>   <glu>
#> 1 KBP1000     South African Reserve Bank l… J      NA      J1      RMILL   KBP1…
#> # … with abbreviated variable names ¹​version_description, ²​frequency,
#> #   ³​unit_of_measure